# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Amiga Specialty HR Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [open, closed, draft] }
      responses:
        "200":
          description: Jobs
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Job" }
    post:
      operationId: createJob
      tags: [jobs]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateJobBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Job" }
  /jobs/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer }
    get:
      operationId: getJob
      tags: [jobs]
      responses:
        "200":
          description: Job
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Job" }
    patch:
      operationId: updateJob
      tags: [jobs]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateJobBody" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Job" }
    delete:
      operationId: deleteJob
      tags: [jobs]
      responses:
        "204": { description: Deleted }
...
  /candidates/{id}/offers:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer }
    get:
      operationId: listOffers
      tags: [offers]
      responses:
        "200":
          description: Offers
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Offer" }
    post:
      operationId: createOffer
      tags: [offers]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateOfferBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Offer" }
  /offers/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer }
    patch:
      operationId: updateOffer
      tags: [offers]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateOfferBody" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Offer" }
    delete:
      operationId: deleteOffer
      tags: [offers]
      responses:
        "204": { description: Deleted }
...
components:
  schemas:
...
    Job:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        department: { type: string }
        location: { type: string }
        employmentType:
          type: string
          enum: [full_time, part_time, contractor]
        salaryMin: { type: ["number", "null"] }
        salaryMax: { type: ["number", "null"] }
        currency: { type: string }
        description: { type: ["string", "null"] }
        status:
          type: string
          enum: [open, closed, draft]
        closingDate: { type: ["string", "null"] }
        hiringManagerId: { type: ["integer", "null"] }
        candidateCount: { type: integer }
        hiredCount: { type: integer }
        createdAt: { type: string }
        updatedAt: { type: string }
      required:
        [id, title, department, location, employmentType, currency, status, candidateCount, hiredCount, createdAt, updatedAt]
    CreateJobBody:
      type: object
      properties:
        title: { type: string }
        department: { type: string }
        location: { type: ["string", "null"] }
        employmentType:
          type: string
          enum: [full_time, part_time, contractor]
        salaryMin: { type: ["number", "null"] }
        salaryMax: { type: ["number", "null"] }
        currency: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        status:
          type: string
          enum: [open, closed, draft]
        closingDate: { type: ["string", "null"] }
        hiringManagerId: { type: ["integer", "null"] }
      required: [title, department, employmentType]
    UpdateJobBody:
      type: object
      properties:
        title: { type: string }
        department: { type: string }
        location: { type: string }
        employmentType:
          type: string
          enum: [full_time, part_time, contractor]
        salaryMin: { type: ["number", "null"] }
        salaryMax: { type: ["number", "null"] }
        currency: { type: string }
        description: { type: ["string", "null"] }
        status:
          type: string
          enum: [open, closed, draft]
        closingDate: { type: ["string", "null"] }
        hiringManagerId: { type: ["integer", "null"] }
...
    Offer:
      type: object
      properties:
        id: { type: integer }
        candidateId: { type: integer }
        salary: { type: number }
        currency: { type: string }
        startDate: { type: string }
        employmentType:
          type: string
          enum: [full_time, part_time, contractor]
        contractType:
          type: string
          enum: [permanent, fixed_term, contract]
        benefitsNote: { type: ["string", "null"] }
        status:
          type: string
          enum: [draft, sent, accepted, declined, withdrawn]
        sentAt: { type: ["string", "null"] }
        respondedAt: { type: ["string", "null"] }
        declinedReason: { type: ["string", "null"] }
        createdAt: { type: string }
      required: [id, candidateId, salary, currency, startDate, employmentType, contractType, status, createdAt]
    CreateOfferBody:
      type: object
      properties:
        salary: { type: number }
        currency: { type: ["string", "null"] }
        startDate: { type: string }
        employmentType:
          type: string
          enum: [full_time, part_time, contractor]
        contractType:
          type: string
          enum: [permanent, fixed_term, contract]
        benefitsNote: { type: ["string", "null"] }
        status:
          type: string
          enum: [draft, sent, accepted, declined, withdrawn]
      required: [salary, startDate, employmentType, contractType]
    UpdateOfferBody:
      type: object
      properties:
        salary: { type: number }
        currency: { type: string }
        startDate: { type: string }
        employmentType:
          type: string
          enum: [full_time, part_time, contractor]
        contractType:
          type: string
          enum: [permanent, fixed_term, contract]
        benefitsNote: { type: ["string", "null"] }
        status:
          type: string
          enum: [draft, sent, accepted, declined, withdrawn]
        declinedReason: { type: ["string", "null"] }
...